Tensyra/requirements.txt
streamlit
firebase-admin
python-dotenv
requests
openai
pandas
plotly
google-auth
oauthlib
fastapi
uvicorn